Add files via upload
This written assignment if first project is developed for me to understand more about Big O Notation (how to calculate and find the running time for each program). 
In the coding assignment, the first problem is to write two approaches of a problem: iterative or recursive, then find running time and add to list of array.